# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Vice Intelligence Platform — GTA Gaming CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
        "429":
          description: Too many attempts
...
  /admin/me:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

  # ─── CMS ARTICLES ────────────────────────────────────────────────────────────
  /admin/articles:
    get:
      operationId: listAdminArticles
      tags: [cms-articles]
      summary: List all articles (admin)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: status, in: query, schema: { type: string, enum: [published, draft, scheduled, archived] } }
        - { name: search, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticleListResponse"
    post:
      operationId: createAdminArticle
      tags: [cms-articles]
      summary: Create article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}:
    get:
      operationId: getAdminArticle
      tags: [cms-articles]
      summary: Get single article (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
        "404":
          description: Not found
    patch:
      operationId: updateAdminArticle
      tags: [cms-articles]
      summary: Update article
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
    delete:
      operationId: deleteAdminArticle
      tags: [cms-articles]
      summary: Delete article permanently
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /admin/articles/{id}/publish:
    post:
      operationId: publishAdminArticle
      tags: [cms-articles]
      summary: Publish article immediately
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}/draft:
    post:
      operationId: draftAdminArticle
      tags: [cms-articles]
      summary: Move article back to draft
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Drafted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}/schedule:
    post:
      operationId: scheduleAdminArticle
      tags: [cms-articles]
      summary: Schedule article for future publish
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "200":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}/archive:
    post:
      operationId: archiveAdminArticle
      tags: [cms-articles]
      summary: Archive article
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}/restore:
    post:
      operationId: restoreAdminArticle
      tags: [cms-articles]
      summary: Restore archived article to draft
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}/duplicate:
    post:
      operationId: duplicateAdminArticle
      tags: [cms-articles]
      summary: Duplicate article
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArticle"
  /admin/articles/{id}/revisions:
    get:
      operationId: getAdminArticleRevisions
      tags: [cms-articles]
      summary: Get article revision history
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Revisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleRevision"
...
components:
  schemas:
...
    AdminArticle:
      type: object
      required: [id, title, slug, status, createdAt]
      properties:
        id:              { type: integer }
        title:           { type: string }
        slug:            { type: string }
        excerpt:         { type: ["string", "null"] }
        content:         { type: ["string", "null"] }
        featuredImage:   { type: ["string", "null"] }
        game:            { type: ["string", "null"] }
        status:          { type: string }
        viewCount:       { type: integer }
        readingTime:     { type: ["integer", "null"] }
        publishedAt:     { type: ["string", "null"] }
        scheduledAt:     { type: ["string", "null"] }
        updatedAt:       { type: ["string", "null"] }
        createdAt:       { type: string }
        seoTitle:        { type: ["string", "null"] }
        metaDescription: { type: ["string", "null"] }
        focusKeyword:    { type: ["string", "null"] }
        canonicalUrl:    { type: ["string", "null"] }
        ogImage:         { type: ["string", "null"] }
        noIndex:         { type: boolean }
        noFollow:        { type: boolean }
        categoryId:      { type: ["integer", "null"] }
        category:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
    AdminArticleListResponse:
      type: object
      required: [articles, total, page, limit]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/AdminArticle"
        total:  { type: integer }
        page:   { type: integer }
        limit:  { type: integer }
    ArticleInput:
      type: object
      required: [title]
      properties:
        title:           { type: string, minLength: 1 }
        slug:            { type: string }
        excerpt:         { type: string }
        content:         { type: string }
        featuredImage:   { type: string }
        game:            { type: string }
        status:          { type: string, enum: [draft, published, scheduled, archived] }
        scheduledAt:     { type: string }
        categoryId:      { type: integer }
        tagIds:          { type: array, items: { type: integer } }
        seoTitle:        { type: string }
        metaDescription: { type: string }
        focusKeyword:    { type: string }
        canonicalUrl:    { type: string }
        ogImage:         { type: string }
        noIndex:         { type: boolean }
        noFollow:        { type: boolean }
    ArticleUpdate:
      type: object
      properties:
        title:           { type: string, minLength: 1 }
        slug:            { type: string }
        excerpt:         { type: string }
        content:         { type: string }
        featuredImage:   { type: string }
        game:            { type: string }
        status:          { type: string, enum: [draft, published, scheduled, archived] }
        scheduledAt:     { type: string }
        categoryId:      { type: integer }
        tagIds:          { type: array, items: { type: integer } }
        seoTitle:        { type: string }
        metaDescription: { type: string }
        focusKeyword:    { type: string }
        canonicalUrl:    { type: string }
        ogImage:         { type: string }
        noIndex:         { type: boolean }
        noFollow:        { type: boolean }
    ArticleRevision:
      type: object
      required: [id, articleId, createdAt]
      properties:
        id:        { type: integer }
        articleId: { type: integer }
        title:     { type: ["string", "null"] }
        content:   { type: ["string", "null"] }
        createdAt: { type: string }
    ScheduleInput:
      type: object
      required: [scheduledAt]
      properties:
        scheduledAt: { type: string }

    # ── CATEGORY SCHEMAS ─────────────────────────────────────────────────────
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:           { type: integer }
        name:         { type: string }
        slug:         { type: string }
        description:  { type: ["string", "null"] }
        color:        { type: ["string", "null"] }
        parentId:     { type: ["integer", "null"] }
        articleCount: { type: integer }
        createdAt:    { type: string }
...
    Tag:
      type: object
      required: [id, name, slug]
      properties:
        id:           { type: integer }
        name:         { type: string }
        slug:         { type: string }
        articleCount: { type: integer }
        createdAt:    { type: string }
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AdminSession:
      type: object
      required: [id, username]
      properties:
        id:       { type: integer }
        username: { type: string }

    # ── MISC SCHEMAS ─────────────────────────────────────────────────────────
...