# lib/api-zod/src/generated/types/articleRevision.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Vice Intelligence Platform — GTA Gaming CMS API
 * OpenAPI spec version: 0.2.0
 */

export interface ArticleRevision {
  id: number;
  articleId: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  content?: string | null;
  createdAt: string;
}